fix: 플레이어 스탯 수정사항 반영

+++ Project_LoL/Assets/_Scripts/LYC/Data/PlayerData.cs
@@ -1,21 +1,21 @@
 ﻿public class PlayerData
 {
-	public int HP;
+	public float HP;
 	public int AtkDamage;
-	public float AtkSpeed;
-	public int MoveSpeed;
+	public int AtkSpeed;
+	public float MoveSpeed;
 	public float CritRate;
 	public int CritDamage;
 	public float DashTime;
 	public float DashCooldown;
 	public float DashInvincTime;
 
-	[field: UnityEngine.SerializeField] public int CurrentHp { get; private set; }
+	[field: UnityEngine.SerializeField] public float CurrentHp { get; private set; }
 	[field: UnityEngine.SerializeField] public int CurrentExp { get; private set; }
 	[field: UnityEngine.SerializeField] public int CurrentLevel { get; private set; }
 	[field: UnityEngine.SerializeField] public int CurrentGold { get; private set; }
 
-	public void SetHp(int value)
+	public void SetHp(float value)
 	{
 		CurrentHp = value;
 	}
